refactor: split report generation modules
